[IA64] GNTMAP_readonly support linux part
add read only mapping support.

+++ linux-2.6-xen-sparse/include/asm-ia64/hypercall.h
@@ -494,7 +494,7 @@ HYPERVISOR_zap_physmap(unsigned long gpfn, unsigned int extent_order)
 
 static inline unsigned long
 __HYPERVISOR_add_physmap(unsigned long gpfn, unsigned long mfn,
-			 unsigned int flags, domid_t domid)
+			 unsigned long flags, domid_t domid)
 {
 	return _hypercall_imm4(unsigned long, ia64_dom0vp_op,
 			       IA64_DOM0VP_add_physmap, gpfn, mfn, flags,
@@ -503,7 +503,7 @@ __HYPERVISOR_add_physmap(unsigned long gpfn, unsigned long mfn,
 
 static inline unsigned long
 HYPERVISOR_add_physmap(unsigned long gpfn, unsigned long mfn,
-		       unsigned int flags, domid_t domid)
+		       unsigned long flags, domid_t domid)
 {
 	unsigned long ret = 0;
 	BUG_ON(!running_on_xen);//XXX
